Add benchmark for neuquant
There were no benchmarks for neuquant, and it's worthwhile to figure out what consumes time when
quantizing the colors. This creates two benchmarks, one for the creation of the quantization,
and one for the searching of the nearest color. It tests on a static image which is mostly green
and red.